+++ pyconnect/clients/kafka.py
@@ -1,11 +1,11 @@
 """
-clients.py
+kafka.py
 
 Client services used to support internal and external transactions.
-Services instances are bound to data attributes and accessed through "get" functions.
+Service instances are bound to data attributes and accessed through "get" functions.
 """
-
-import ssl
+import httpx
+import json
 import logging
 from asyncio import (get_event_loop,
                      get_running_loop)
@@ -13,20 +13,28 @@ from confluent_kafka import (Producer,
                              Consumer,
                              KafkaException,
                              KafkaError,
+                             Message,
                              TopicPartition)
-from nats.aio.client import Client as NatsClient
-from threading import Thread
-from pyconnect.config import get_settings
-from typing import Optional
+from confluent_kafka.admin import (AdminClient,
+                                   NewTopic)
+from pyconnect.config import (get_settings,
+                              kafka_sync_topic)
 from pyconnect.exceptions import KafkaMessageNotFoundError
+from pyconnect.support.encoding import decode_to_dict
+from threading import Thread
+from typing import (Callable,
+                    List,
+                    Optional)
+
 
 logger = logging.getLogger(__name__)
-
 # client instances
 kafka_producer = None
-nats_client = None
+kafka_listeners = []
 
-
+# ******************************************
+# Confluent async Kafka producer and methods
+# ******************************************
 class ConfluentAsyncKafkaProducer:
     """
     Confluent's AsyncIO Wrapper for a Kafka Producer
@@ -83,7 +91,6 @@ class ConfluentAsyncKafkaProducer:
         self._producer.produce(topic, value, on_delivery=ack)
         return result
 
-
 def get_kafka_producer() -> Optional[ConfluentAsyncKafkaProducer]:
     """
     :return: a connected ConfluentAsyncKafkaProducer instance
@@ -100,6 +107,9 @@ def get_kafka_producer() -> Optional[ConfluentAsyncKafkaProducer]:
     return kafka_producer
 
 
+# ******************************************
+# Confluent async Kafka consumer and methods
+# ******************************************
 class ConfluentAsyncKafkaConsumer:
     """
     A high-level KafkaConsumer class in order to create mutliple instances of
@@ -234,25 +244,139 @@ def get_kafka_consumer(topic_name: str,
     return kafka_consumer
 
 
-async def get_nats_client() -> Optional[NatsClient]:
+# ************************************************
+# Confluent async Kafka topic listener and methods
+# ************************************************
+class ConfluentAsyncKafkaListener:
     """
-    :return: a connected NATS client instance
+    Confluent's AsyncIO Wrapper for a Kafka topic listener
+    Adapted from https://github.com/confluentinc/confluent-kafka-python
     """
-    global nats_client
+    def __init__(self, configs, loop=None):
+        self.topics = None
+        self.callback = None
+        self._loop = loop or get_event_loop()
+        self._consumer = Consumer(configs)
+        self._cancelled = False
+        self._poll_thread = Thread(target=self._poll_loop)
+        self._poll_thread.start()
 
-    if not nats_client:
-        settings = get_settings()
+    def _poll_loop(self):
+        while not self._cancelled:
+            if self.topics and self.callback:
+                msg = self._consumer.poll(timeout=1.0)
+                if msg is None: continue
 
-        ssl_ctx = ssl.create_default_context(purpose=ssl.Purpose.SERVER_AUTH)
-        ssl_ctx.load_verify_locations(settings.nats_rootCA_file)
-        ssl_ctx.load_cert_chain(settings.nats_cert_file, settings.nats_key_file)
+                if msg.error():
+                    if msg.error().code() == KafkaError._PARTITION_EOF:
+                        # End of partition event
+                        logger.debug(f'Listener: {msg.topic()} [{msg.partition()}] reached end, offset {msg.offset()}')
+                    elif msg.error():
+                        raise KafkaException(msg.error())
+                else:
+                    self.callback(msg)
 
-        nats_client = NatsClient()
-        await nats_client.connect(
-            servers=settings.nats_servers,
-            loop=get_running_loop(),
-            tls=ssl_ctx,
-            allow_reconnect=settings.nats_allow_reconnect,
-            max_reconnect_attempts=settings.nats_max_reconnect_attempts)
+    def close(self):
+        self._cancelled = True
+        self.topics = None
+        self.callback = None
+        self._poll_thread.join()
 
-    return nats_client
+    def listen(self, topics: List[str], callback: Callable):
+        self._consumer.subscribe(topics)
+        self.topics = topics
+        self.callback = callback
+
+
+def create_kafka_listeners():
+    """
+    Create an instance of each Kafka listener.  Add additional listeners as needed.
+    """
+    listener = start_sync_event_listener()
+    kafka_listeners.append(listener)
+
+
+def start_sync_event_listener():
+    """
+    Listen on the Kafka kafka_sync_topic for NATS sync messages.  Create the topic if it doesn't exist.
+    """
+    settings = get_settings()
+    client = AdminClient({'bootstrap.servers': ''.join(settings.kafka_bootstrap_servers)})
+    metadata = client.list_topics()
+    if metadata.topics.get(kafka_sync_topic) is None:
+        new_topic = NewTopic(kafka_sync_topic,
+                             num_partitions=settings.kafka_admin_new_topic_partitions,
+                             replication_factor=settings.kafka_admin_new_topic_replication_factor)
+        client.create_topics([new_topic])
+        logger.debug(f'start_sync_event_listener: created topic = {kafka_sync_topic}')
+
+    kafka_listener = get_kafka_listener()
+    kafka_listener.listen([kafka_sync_topic], lfh_sync_msg_handler)
+    return kafka_listener
+
+
+def lfh_sync_msg_handler(msg: Message):
+    """
+    Process NATS synchronization messages stored in Kafka in kafka_sync_topic
+    """
+    settings = get_settings()
+    logger.debug(f'lfh_sync_msg_handler: received message = {msg.value()}from topic={msg.topic()}')
+    msg_str = msg.value().decode()
+    message = json.loads(msg_str)
+
+    data_encoded = message['data']
+    data = decode_to_dict(data_encoded)
+    origin_url = message['consuming_endpoint_url']
+    destination_url = 'https://localhost:'+str(settings.uvicorn_port)+origin_url
+
+    result = httpx.post(destination_url, json=data, verify=settings.certificate_verify)
+    logger.debug(f'lfh_sync_msg_handler: posted message to {destination_url}  result = {result}')
+
+
+def remove_kafka_listeners():
+    """
+    Stop all Kafka listeners
+    """
+    for listener in kafka_listeners:
+        listener.close()
+
+
+def get_kafka_listener() -> Optional[ConfluentAsyncKafkaListener]:
+    """
+    :return: a new connected ConfluentAsyncKafkaListener instance
+    """
+    settings = get_settings()
+    listener_config = {
+        'bootstrap.servers': ''.join(settings.kafka_bootstrap_servers),
+        'group.id': settings.kafka_consumer_default_group_id,
+        'enable.auto.commit': settings.kafka_consumer_default_enable_auto_commit
+    }
+    return ConfluentAsyncKafkaListener(configs=listener_config,
+                                       loop=get_running_loop())
+
+
+# *************************************
+# Kafka callback for storage operations
+# *************************************
+class KafkaCallback():
+    """
+    Store returned data from the Kafka callback
+    """
+    kafka_status = None
+    kafka_result = None
+
+    def get_kafka_result(self, err: object, msg: object):
+        """
+        Kafka producer callback for storage operations
+        :param err: If error, the error returned from Kafka.
+        :param msg: If success, the topic, partition and offset of the stored message.
+        """
+        if err is not None:
+            self.kafka_status = 'error'
+            logging.debug(self.kafka_status)
+            raise KafkaStorageError(f'Failed to deliver message: {str(msg)} {str(err)}')
+        else:
+            self.kafka_status = 'success'
+            self.kafka_result = f'{msg.topic()}:{msg.partition()}:{msg.offset()}'
+            logging.debug(f'Produced record to topic {msg.topic()} ' \
+                          f'partition [{msg.partition()}] @ offset {msg.offset()}')
