+++ pyconnect/routes/data.py
@@ -9,7 +9,7 @@ from pydantic import (BaseModel,
 from fastapi.routing import (APIRouter,
                              HTTPException)
 from typing import Optional
-from pyconnect.clients import get_kafka_consumer
+from pyconnect.clients.kafka import get_kafka_consumer
 from pyconnect.exceptions import KafkaMessageNotFoundError
 from confluent_kafka import KafkaException
