+++ tests/routes/test_fhir.py
@@ -2,12 +2,12 @@
 test_fhir.py
 Tests the /fhir endpoint
 """
+import asyncio
 import pytest
-from pyconnect import clients
+from pyconnect.clients import kafka
 from pyconnect.config import get_settings
 from pyconnect.workflows.fhir import FhirWorkflow
 from starlette.responses import Response
-import asyncio
 from unittest.mock import AsyncMock
 
 
@@ -180,7 +180,7 @@ async def test_fhir_post(async_test_client,
     :param settings: pyConnect configuration settings fixture
     """
     with monkeypatch.context() as m:
-        m.setattr(clients, 'ConfluentAsyncKafkaProducer', mock_async_kafka_producer)
+        m.setattr(kafka, 'ConfluentAsyncKafkaProducer', mock_async_kafka_producer)
         m.setattr(FhirWorkflow, 'synchronize', AsyncMock())
 
         async with async_test_client as ac:
@@ -233,7 +233,7 @@ async def test_fhir_post_with_transmit(async_test_client,
         self.use_response = True
 
     with monkeypatch.context() as m:
-        m.setattr(clients, 'ConfluentAsyncKafkaProducer', mock_async_kafka_producer)
+        m.setattr(kafka, 'ConfluentAsyncKafkaProducer', mock_async_kafka_producer)
         m.setattr(FhirWorkflow, 'transmit', mock_workflow_transmit)
         m.setattr(FhirWorkflow, 'synchronize', AsyncMock())
 
@@ -261,7 +261,7 @@ async def test_fhir_post_endpoints(async_test_client,
     :param settings: pyConnect configuration settings fixture
     """
     with monkeypatch.context() as m:
-        m.setattr(clients, 'ConfluentAsyncKafkaProducer', mock_async_kafka_producer)
+        m.setattr(kafka, 'ConfluentAsyncKafkaProducer', mock_async_kafka_producer)
         m.setattr(FhirWorkflow, 'synchronize', AsyncMock())
 
         async with async_test_client as ac:
